backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
whitenoise==6.7.0
celery==5.4.0
redis==5.0.4
psycopg[binary]==3.1.19
dj-database-url==2.2.0
gunicorn==22.0.0
pytest==8.2.1
pytest-django==4.8.0
